7709 - Migrate all CosmosDb Test project files to CQRS
