+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.API.sln
@@ -65,8 +65,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.csproj", "{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests.csproj", "{688708D6-CF24-452C-BBFD-59161F982AA0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions.csproj", "{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Configuration", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Configuration\xxENSONOxx.xxSTACKSxx.Shared.Configuration.csproj", "{BFE98A28-8B88-4CC6-BECF-324D0D4CD760}"
@@ -397,18 +395,6 @@ Global
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|x64.Build.0 = Release|Any CPU
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|x86.ActiveCfg = Release|Any CPU
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|x86.Build.0 = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x64.Build.0 = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x86.Build.0 = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x64.ActiveCfg = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x64.Build.0 = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x86.ActiveCfg = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x86.Build.0 = Release|Any CPU
 		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7BEFAEE8-4F24-434B-ABEA-4E904804266C}.Debug|x64.ActiveCfg = Debug|Any CPU
