+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure.IntegrationTests/CosmosDb/Integration/IntegrationCollection.cs
@@ -1,0 +1,10 @@
+﻿using Xunit;
+
+namespace xxENSONOxx.xxSTACKSxx.Infrastructure.IntegrationTests.CosmosDb.Integration;
+
+[CollectionDefinition(Name)]
+public class IntegrationCollection : ICollectionFixture<IntegrationFixture>
+{
+    // Must be unique compared with other collections within the project
+    public const string Name = "IntegrationFixture";
+}
