+++ src/func-cosmosdb-worker/.template.config/template.json
@@ -30,7 +30,6 @@
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Core.Tests/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.Abstractions/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB/**",
-				"shared/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.DependencyInjection/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus/**",
 				"shared/xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.ServiceBus.Tests/**",
