+++ src/func-cosmosdb-worker/src/functions/xxENSONOxx.xxSTACKSxx.Worker.sln
@@ -29,8 +29,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.csproj", "{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests.csproj", "{688708D6-CF24-452C-BBFD-59161F982AA0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions.csproj", "{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Configuration", "..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Configuration\xxENSONOxx.xxSTACKSxx.Shared.Configuration.csproj", "{BFE98A28-8B88-4CC6-BECF-324D0D4CD760}"
@@ -95,10 +93,6 @@ Global
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -133,7 +127,6 @@ Global
 		{3F133394-CF28-4110-BA3D-EBF404208A81} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{06DE9A76-034B-4975-8365-A7F88EF7954E} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
-		{688708D6-CF24-452C-BBFD-59161F982AA0} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{BFE98A28-8B88-4CC6-BECF-324D0D4CD760} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
 		{80757BB7-48FF-4EE0-9CBF-AF49B20D07E4} = {1F2AAD73-889C-4901-92F9-646EFF6234B4}
