--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests/appsettings.json
@@ -1,11 +1,0 @@
-{
-  "CosmosDb": {
-    "DatabaseAccountUri": "https://localhost:8081/",
-    "DatabaseName": "Stacks",
-    "PrimaryKey": "",
-    "SecurityKeySecret": {
-      "Identifier": "COSMOSDB_KEY",
-      "Source": "Environment"
-    }
-  }
-}
