+++ xxENSONOxx.xxSTACKSxx.sln
@@ -59,8 +59,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Share
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.csproj", "{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests\xxENSONOxx.xxSTACKSxx.Shared.Data.Documents.CosmosDB.Tests.csproj", "{688708D6-CF24-452C-BBFD-59161F982AA0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions.csproj", "{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Configuration", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Configuration\xxENSONOxx.xxSTACKSxx.Shared.Configuration.csproj", "{BFE98A28-8B88-4CC6-BECF-324D0D4CD760}"
@@ -449,18 +447,6 @@ Global
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|x64.Build.0 = Release|Any CPU
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|x86.ActiveCfg = Release|Any CPU
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6}.Release|x86.Build.0 = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x64.Build.0 = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Debug|x86.Build.0 = Debug|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x64.ActiveCfg = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x64.Build.0 = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x86.ActiveCfg = Release|Any CPU
-		{688708D6-CF24-452C-BBFD-59161F982AA0}.Release|x86.Build.0 = Release|Any CPU
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -733,7 +719,6 @@ Global
 		{3F133394-CF28-4110-BA3D-EBF404208A81} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{06DE9A76-034B-4975-8365-A7F88EF7954E} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{4856749F-880C-4F3E-B9C9-4E79ADAD09A6} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
-		{688708D6-CF24-452C-BBFD-59161F982AA0} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{AF0ACDAF-CE5D-4D1E-BD7F-D05397931B90} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{BFE98A28-8B88-4CC6-BECF-324D0D4CD760} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{80757BB7-48FF-4EE0-9CBF-AF49B20D07E4} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
